requirements.txt
azure-identity
azure-ai-projects>=2.0.0b1
python-dotenv
streamlit
